Add Makefile for easy distribution
This also starts maintaining the version in `metadata.json` even though
it is discouraged by EGO (extensions.gnome.org). This is because I
expect the extension to be rejected because of the hard requirement to
use the `gdm` session mode. Hence I must maintain the version manually
to ensure no version shenanigans happen.